refactor : 정적리소스 디폴트콘트롤러로 추출
